+++ MT4Expander.vcproj
@@ -380,10 +380,6 @@
 					>
 				</File>
 				<File
-					RelativePath=".\header\lib\win32.h"
-					>
-				</File>
-				<File
 					RelativePath=".\header\lib\window.h"
 					>
 				</File>
@@ -1014,32 +1010,8 @@
 					</FileConfiguration>
 				</File>
 				<File
-					RelativePath=".\src\lib\win32.cpp"
+					RelativePath=".\src\lib\window.cpp"
 					>
-					<FileConfiguration
-						Name="Debug|Win32"
-						>
-						<Tool
-							Name="VCCLCompilerTool"
-							ObjectFile="$(IntDir)\lib\"
-						/>
-					</FileConfiguration>
-					<FileConfiguration
-						Name="Release|Win32"
-						>
-						<Tool
-							Name="VCCLCompilerTool"
-							ObjectFile="$(IntDir)\lib\"
-						/>
-					</FileConfiguration>
-					<FileConfiguration
-						Name="Release (private)|Win32"
-						>
-						<Tool
-							Name="VCCLCompilerTool"
-							ObjectFile="$(IntDir)\lib\"
-						/>
-					</FileConfiguration>
 				</File>
 				<Filter
 					Name="ui"
