+++ header/lib/format.h
@@ -1,5 +1,4 @@
 #pragma once
 #include "expander.h"
 
-
 char* WINAPI NumberFormat(double value, const char* format);
