+++ header/lib/memory.h
@@ -2,7 +2,6 @@
 #include "stdafx.h"
 #include "struct/mt4/MqlString.h"
 
-
 uint WINAPI GetBoolsAddress(const BOOL values[]);
 uint WINAPI GetDoublesAddress(const double values[]);
 uint WINAPI GetIntsAddress(const int values[]);
