--- header/lib/win32.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "expander.h"
-
-int     WINAPI EnumChildWindowsToDebug(HWND hWnd, BOOL recursive = FALSE);
-char*   WINAPI GetClassNameA(HWND hWnd);
-wchar*  WINAPI GetClassNameW(HWND hWnd);
-string  WINAPI getClassNameA(HWND hWnd);
-wstring WINAPI getClassNameW(HWND hWnd);
