+++ header/struct/mt4/SymbolSelected.h
@@ -1,7 +1,6 @@
 #pragma once
 #include "expander.h"
 
-
 #pragma pack(push, 1)
 /**
  * MT4 struct SYMBOL_SELECTED (file format of "<data-directory>/history/<trade-server>/symbols.sel")
