+++ header/struct/mt4/Tick.h
@@ -1,7 +1,6 @@
 #pragma once
 #include "expander.h"
 
-
 #pragma pack(push, 1)
 /**
  * MT4 struct TICK (file format of "<data-directory>/history/<trade-server>/ticks.raw")
