+++ header/struct/win32/ProcessInformation.h
@@ -1,6 +1,5 @@
 #pragma once
-#include "stdafx.h"
-
+#include "expander.h"
 
 /**
  * Win32 struct PROCESS_INFORMATION
