+++ header/struct/win32/StartupInfo.h
@@ -1,6 +1,5 @@
 #pragma once
-#include "stdafx.h"
-
+#include "expander.h"
 
 /**
  * Win32 struct STARTUPINFOA (ANSI-Version)
