+++ src/lib/array.cpp
@@ -3,7 +3,7 @@
 
 
 /**
- * Initialize a range of array elements with a custom value.
+ * Initialize an array range with a custom value.
  *
  * @param  T   array[]          - array
  * @param  int size             - number of elements in the array
@@ -15,23 +15,23 @@
  */
 template <typename T>
 BOOL WINAPI InitializeArray(T array[], int size, T initValue, int from, int count/*=INT_MAX*/) {
-   if ((uint)array < MIN_VALID_POINTER) return(!error(ERR_INVALID_PARAMETER, "invalid parameter array: 0x%p (not a valid pointer)", array));
-   if (size < 0)                        return(!error(ERR_INVALID_PARAMETER, "invalid parameter size: %d (must be >= 0)", size));
-   if (from < 0)                        return(!error(ERR_INVALID_PARAMETER, "invalid parameter from: %d (must be >= 0)", from));
-   if (from >= size)                    return(!error(ERR_INVALID_PARAMETER, "invalid parameter from: %d (out of range)", from));
-   if (count < 0)                       return(!error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (must be >= 0)", count));
+   if ((uint)array < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter array: 0x%p (not a valid pointer)", array);
+   if (size < 0)                        return !error(ERR_INVALID_PARAMETER, "invalid parameter size: %d (must be >= 0)", size);
+   if (from < 0)                        return !error(ERR_INVALID_PARAMETER, "invalid parameter from: %d (must be >= 0)", from);
+   if (from >= size)                    return !error(ERR_INVALID_PARAMETER, "invalid parameter from: %d (out of range)", from);
+   if (count < 0)                       return !error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (must be >= 0)", count);
 
    if (count == INT_MAX) {
       count = size - from;
    }
    else {
       // @see  https://stackoverflow.com/questions/3944505/detecting-signed-overflow-in-c-c#3947943
-      if (count > INT_MAX-from) return(!error(ERR_INVALID_PARAMETER, "integer overflow detected (from: %d, count: %d)", from, count));
-      if (from+count > size)    return(!error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (out of range)", count));
+      if (count > INT_MAX-from) return !error(ERR_INVALID_PARAMETER, "integer overflow detected (from: %d, count: %d)", from, count);
+      if (from+count > size)    return !error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (out of range)", count);
    }
 
    std::fill_n(&array[from], count, initValue);
-   return(TRUE);
+   return TRUE;
 }
 
 
@@ -46,7 +46,7 @@ template BOOL InitializeArray<double>(double[], int, double, int, int);
 
 
 /**
- * Initialize a range of <MQL4.0:bool> array elements with a custom value.
+ * Initialize an array<MQL4.0::bool> range with a custom value.
  *
  * @param  BOOL array[]   - array with sizeof(element) = 4
  * @param  int  size      - number of elements in the array
@@ -57,13 +57,13 @@ template BOOL InitializeArray<double>(double[], int, double, int, int);
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeBOOLArray(BOOL array[], int size, BOOL initValue, int from, int count) {
-   return(InitializeArray(array, size, (int)(initValue!=0), from, count));
+   return InitializeArray(array, size, (int)(initValue!=0), from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <MQL4.5:bool> array elements with a custom value.
+ * Initialize an array<MQL4.5::bool> range with a custom value.
  *
  * @param  bool array[]   - array with sizeof(element) = 1
  * @param  int  size      - number of elements in the array
@@ -74,13 +74,13 @@ BOOL WINAPI InitializeBOOLArray(BOOL array[], int size, BOOL initValue, int from
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeBoolArray(bool array[], int size, bool initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <char> array elements with a custom value.
+ * Initialize an array<MQL4.5::char> range with a custom value.
  *
  * @param  char array[]   - array
  * @param  int  size      - number of elements in the array
@@ -91,13 +91,13 @@ BOOL WINAPI InitializeBoolArray(bool array[], int size, bool initValue, int from
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeCharArray(char array[], int size, char initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <short> array elements with a custom value.
+ * Initialize an array<MQL4.5::short> range with a custom value.
  *
  * @param  short array[]   - array
  * @param  int   size      - number of elements in the array
@@ -108,13 +108,13 @@ BOOL WINAPI InitializeCharArray(char array[], int size, char initValue, int from
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeShortArray(short array[], int size, short initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <int> array elements with a custom value.
+ * Initialize an array<int> range with a custom value.
  *
  * @param  int array[]   - array
  * @param  int size      - number of elements in the array
@@ -125,13 +125,13 @@ BOOL WINAPI InitializeShortArray(short array[], int size, short initValue, int f
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeIntArray(int array[], int size, int initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <MQL4.5:long> array elements with a custom value.
+ * Initialize an array<MQL4.5::long> range with a custom value.
  *
  * @param  int64 array[]   - array with sizeof(element) = 8
  * @param  int   size      - number of elements in the array
@@ -142,13 +142,13 @@ BOOL WINAPI InitializeIntArray(int array[], int size, int initValue, int from, i
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeLongArray(int64 array[], int size, int64 initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <float> array elements with a custom value.
+ * Initialize an array<MQL4.5::float> range with a custom value.
  *
  * @param  float  array[]   - array
  * @param  int    size      - number of elements in the array
@@ -159,13 +159,13 @@ BOOL WINAPI InitializeLongArray(int64 array[], int size, int64 initValue, int fr
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeFloatArray(float array[], int size, float initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Initialize a range of <double> array elements with a custom value.
+ * Initialize an array<double> range with a custom value.
  *
  * @param  double array[]   - array
  * @param  int    size      - number of elements in the array
@@ -176,13 +176,13 @@ BOOL WINAPI InitializeFloatArray(float array[], int size, float initValue, int f
  * @return BOOL - success status
  */
 BOOL WINAPI InitializeDoubleArray(double array[], int size, double initValue, int from, int count) {
-   return(InitializeArray(array, size, initValue, from, count));
+   return InitializeArray(array, size, initValue, from, count);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Search an <int> array for a value and return its index.
+ * Search an array<int> for a value and return its index.
  *
  * @param  int  array[]                   - array to search
  * @param  int  size                      - number of elements in the array
@@ -193,7 +193,7 @@ BOOL WINAPI InitializeDoubleArray(double array[], int size, double initValue, in
  *               -2 in case of errors
  */
 int WINAPI SearchIntArray(const int array[], int size, int value, BOOL reverseIndexed = FALSE) {
-   return(EMPTY);
+   return EMPTY;
    //#pragma EXPANDER_EXPORT
 }
 
@@ -211,17 +211,17 @@ int WINAPI SearchIntArray(const int array[], int size, int value, BOOL reverseIn
  */
 template <typename T>
 BOOL WINAPI ShiftIndicatorBuffer(T buffer[], int size, int count, T emptyValue) {
-   if ((uint)buffer < MIN_VALID_POINTER) return(!error(ERR_INVALID_PARAMETER, "invalid parameter buffer: 0x%p (not a valid pointer)", buffer));
-   if (size < 0)                         return(!error(ERR_INVALID_PARAMETER, "invalid parameter size: %d (must be >= 0)", size));
-   if (count < 0)                        return(!error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (must be >= 0)", count));
-   if (count > size)                     return(!error(ERR_INVALID_PARAMETER, "invalid parameter count=%d for size=%d (out of range)", count, size));
-   if (!size || !count) return(TRUE);
+   if ((uint)buffer < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter buffer: 0x%p (not a valid pointer)", buffer);
+   if (size < 0)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter size: %d (must be >= 0)", size);
+   if (count < 0)                        return !error(ERR_INVALID_PARAMETER, "invalid parameter count: %d (must be >= 0)", count);
+   if (count > size)                     return !error(ERR_INVALID_PARAMETER, "invalid parameter count=%d for size=%d (out of range)", count, size);
+   if (!size || !count) return TRUE;
 
    if (count < size) {
-      MoveMemory((void*)&buffer[0], &buffer[count], (size-count)*sizeof(buffer[0]));
+      MoveMemory((void*)&buffer[0], &buffer[count], (size-count) * sizeof(buffer[0]));
    }
    std::fill_n(&buffer[size-count], count, emptyValue);
-   return(TRUE);
+   return TRUE;
 }
 
 
@@ -236,7 +236,7 @@ template BOOL ShiftIndicatorBuffer<double>(double[], int, int, double);
 
 
 /**
- * Shift the content of an <MQL4.0:bool> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of an <MQL4.0::bool> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  BOOL buffer[]   - timeseries array with sizeof(element) = 4
@@ -247,13 +247,13 @@ template BOOL ShiftIndicatorBuffer<double>(double[], int, int, double);
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftBOOLIndicatorBuffer(BOOL buffer[], int size, int count, BOOL emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, (int)(emptyValue!=0)));
+   return ShiftIndicatorBuffer(buffer, size, count, (int)(emptyValue != 0));
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Shift the content of an <MQL4.5:bool> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of an <MQL4.5::bool> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  bool buffer[]   - timeseries array with sizeof(element) = 1
@@ -264,13 +264,13 @@ BOOL WINAPI ShiftBOOLIndicatorBuffer(BOOL buffer[], int size, int count, BOOL em
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftBoolIndicatorBuffer(bool buffer[], int size, int count, bool emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Shift the content of a <char> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of a <MQL4.5::char> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  char buffer[]   - timeseries array
@@ -281,13 +281,13 @@ BOOL WINAPI ShiftBoolIndicatorBuffer(bool buffer[], int size, int count, bool em
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftCharIndicatorBuffer(char buffer[], int size, int count, char emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Shift the content of a <short> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of a <MQL4.5::short> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  short buffer[]   - timeseries array
@@ -298,7 +298,7 @@ BOOL WINAPI ShiftCharIndicatorBuffer(char buffer[], int size, int count, char em
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftShortIndicatorBuffer(short buffer[], int size, int count, short emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
@@ -315,13 +315,13 @@ BOOL WINAPI ShiftShortIndicatorBuffer(short buffer[], int size, int count, short
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftIntIndicatorBuffer(int buffer[], int size, int count, int emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Shift the content of an <MQL4.5:long> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of an <MQL4.5::long> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  int64 buffer[]   - timeseries array with sizeof(element) = 8
@@ -332,13 +332,13 @@ BOOL WINAPI ShiftIntIndicatorBuffer(int buffer[], int size, int count, int empty
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftLongIndicatorBuffer(int64 buffer[], int size, int count, int64 emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
 
 /**
- * Shift the content of a <float> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
+ * Shift the content of a <MQL4.5::float> timeseries array (e.g. an indicator buffer) by the specified number of elements. Discards
  * elements at the beginning (the oldest values of a timeseries).
  *
  * @param  float  buffer[]   - timeseries array
@@ -349,7 +349,7 @@ BOOL WINAPI ShiftLongIndicatorBuffer(int64 buffer[], int size, int count, int64 
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftFloatIndicatorBuffer(float buffer[], int size, int count, float emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
 
@@ -366,6 +366,6 @@ BOOL WINAPI ShiftFloatIndicatorBuffer(float buffer[], int size, int count, float
  * @return BOOL - success status
  */
 BOOL WINAPI ShiftDoubleIndicatorBuffer(double buffer[], int size, int count, double emptyValue) {
-   return(ShiftIndicatorBuffer(buffer, size, count, emptyValue));
+   return ShiftIndicatorBuffer(buffer, size, count, emptyValue);
    #pragma EXPANDER_EXPORT
 }
