+++ src/lib/format.cpp
@@ -8,14 +8,15 @@
  * @param  doube value
  * @param  char* format - format control string as used for printf()
  *
- * @return char* - formatted string or NULL in case of errors
+ * @return char* - formatted string or a NULL pointer in case of errors
  *
+ * Format codes:
  * @see  https://alvinalexander.com/programming/printf-format-cheat-sheet
  * @see  http://www.cplusplus.com/reference/cstdio/printf/
  * @see  ms-help://MS.VSCC.v90/MS.MSDNQTR.v90.en/dv_vccrt/html/664b1717-2760-4c61-bd9c-22eee618d825.htm
  */
 char* WINAPI NumberFormat(double value, const char* format) {
-   if (format && (uint)format < MIN_VALID_POINTER) return((char*)!error(ERR_INVALID_PARAMETER, "invalid parameter format: 0x%p (not a valid pointer)", format));
+   if (format && (uint)format < MIN_VALID_POINTER) return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter format: 0x%p (not a valid pointer)", format);
 
    return asformat(format, value);           // caller must free()
    #pragma EXPANDER_EXPORT
