+++ src/lib/helper.cpp
@@ -2,15 +2,6 @@
 #include "lib/helper.h"
 #include "lib/string.h"
 #include "lib/terminal.h"
-#include <map>
-
-typedef std::map<string, int>    IntegerMap;       // map = associative array
-typedef std::map<string, double> DoubleMap;        // new elements are added "by-copy"
-typedef std::map<string, string> StringMap;        //
-
-IntegerMap g_intWndProperties;                     // a map with integers stored as window properties
-DoubleMap  g_doubleWndProperties;                  // a map with doubles stored as window properties
-StringMap  g_stringWndProperties;                  // a map with strings stored as window properties
 
 
 /**
@@ -34,125 +25,6 @@ BOOL WINAPI IsDebugBuild() {
  */
 DWORD WINAPI GetLastWin32Error() {
    return GetLastError();
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function reads the text directly from the window structure, it doesn't send a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return char* - text or a NULL pointer in case of errors
- */
-char* WINAPI GetInternalWindowTextA(HWND hWnd) {
-   SetLastError(NO_ERROR);
-   string text = getInternalWindowTextA(hWnd);
-   if (GetLastError()) return NULL;
-
-   return sdup(text.c_str());                            // caller must free()
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function reads the text directly from the window structure, it doesn't send a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return string - text or an empty string in case of errors; call GetLastError() for details
- */
-string WINAPI getInternalWindowTextA(HWND hWnd) {
-   wstring text = getInternalWindowTextW(hWnd);
-   return utf16ToAnsi(text);
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function reads the text directly from the window structure, it doesn't send a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return wchar* - text or a NULL pointer in case of errors
- */
-wchar* WINAPI GetInternalWindowTextW(HWND hWnd) {
-   SetLastError(NO_ERROR);
-   wstring text = getInternalWindowTextW(hWnd);
-   if (GetLastError()) return NULL;
-
-   return wsdup(text.c_str());                           // caller must free()
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function reads the text directly from the window structure, it doesn't send a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return wstring - text or an empty string in case of errors; call GetLastError() for details
- */
-wstring WINAPI getInternalWindowTextW(HWND hWnd) {
-   wchar* buffer = NULL;
-   int chars = 64, copiedChars = chars;
-
-   SetLastError(NO_ERROR);
-   while (copiedChars >= chars-1) {                      // if (length == chars-1) the string may have been truncated
-      chars <<= 1;                                       // double the length (starts with 128 chars)
-      buffer = (wchar*) alloca(chars * sizeof(wchar));
-      copiedChars = InternalGetWindowText(hWnd, buffer, chars);
-   }
-   if (!copiedChars && GetLastError()) return _empty_wstr(error(ERR_WIN32_ERROR + GetLastError(), "InternalGetWindowText()"));
-
-   return wstring(buffer);
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function gets the text as a response to a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return char* - text (may be empty) or a NULL pointer in case of errors
- */
-char* WINAPI GetWindowTextA(HWND hWnd) {
-   wchar* utf16Text = GetWindowTextW(hWnd);
-   if (!utf16Text) return NULL;
-
-   char* ansiText = utf16ToAnsi(utf16Text);
-   free(utf16Text);
-
-   return ansiText;                                      // caller must free()
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the text of the specified window's title bar. If the window is a control the text of the control is obtained.
- * This function gets the text as a response to a WM_GETTEXT message.
- *
- * @param  HWND hWnd - window handle
- *
- * @return wchar* - text (may be empty) or a NULL pointer in case of errors
- */
-wchar* WINAPI GetWindowTextW(HWND hWnd) {
-   wchar* buffer = NULL;
-   int chars = 64, copiedChars = chars;
-
-   SetLastError(NO_ERROR);
-   while (copiedChars >= chars-1) {                      // if (length == chars-1) the string may have been truncated
-      chars <<= 1;                                       // double the size (starts with 128 chars)
-      buffer = (wchar*) alloca(chars * sizeof(wchar));
-      copiedChars = GetWindowTextW(hWnd, buffer, chars);
-   }
-   if (!copiedChars && GetLastError()) return (wchar*)!error(ERR_WIN32_ERROR + GetLastError(), "GetWindowTextW()");
-
-   return wsdup(buffer);                                 // caller must free()
    #pragma EXPANDER_EXPORT
 }
 
@@ -327,348 +199,6 @@ BOOL WINAPI IsWindowAreaVisible(HWND hWnd) {
 
    return (region != NULLREGION);
    #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Alias of user32::GetPropA()
- *
- * Retrieves a named value stored in the property list of the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - property name
- *
- * @return HANDLE - the stored value or NULL if no such property exists in the window
- */
-HANDLE WINAPI GetWindowPropertyA(HWND hWnd, const char* name) {
-   return GetPropA(hWnd, name);
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Alias of user32::SetPropA()
- *
- * Sets a named value stored in the property list of the specified window.
- *
- * @param  HWND   hWnd  - window handle
- * @param  char*  name  - property name
- * @param  HANDLE value - property value
- *
- * @return BOOL - success status
- */
-BOOL WINAPI SetWindowPropertyA(HWND hWnd, const char* name, HANDLE value) {
-   BOOL success = SetPropA(hWnd, name, value);
-   if (!success) {
-      error(ERR_WIN32_ERROR + GetLastError(), "SetPropA()  name=\"%s\"", name);
-   }
-   return success;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Alias of user32::RemovePropA()
- *
- * Removes a named value from the property list of the specified window.
- *
- * @param  HWND   hWnd  - window handle
- * @param  char*  name  - property name
- *
- * @return HANDLE - the removed value or NULL if no such property exists in the window
- */
-HANDLE WINAPI RemoveWindowPropertyA(HWND hWnd, const char* name) {
-   return RemovePropA(hWnd, name);
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Callback function for EnumWindowPropertiesA(). Prints property names/values to the debug output.
- *
- * @param  HWND      hWnd   - window whose property list is being enumerated
- * @param  wchar*    name   - property name
- * @param  HANDLE    value  - property value
- * @param  ULONG_PTR prefix - limiting name prefix of properties to print
- *
- * @return BOOL - whether to continue enumeration with the next property
- */
-BOOL CALLBACK EnumWindowPropertiesProcW(HWND hwnd, wchar* name, HANDLE value, ULONG_PTR prefix) {
-   wchar* wPrefix = (wchar*)prefix;
-   if (wPrefix && !*wPrefix) {
-      wPrefix = NULL;
-   }
-   if (wPrefix && !StrStartsWith(name, wPrefix)) {
-      return TRUE;
-   }
-   debug("property %S = %d", name, value);
-   return TRUE;
-}
-
-
-/**
- * Enumerates the properties of the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - limiting name prefix of properties to print
- *
- * @return BOOL - success status; FALSE if the function did not find any window properties or in case of errors
- */
-BOOL WINAPI EnumWindowPropertiesA(HWND hWnd, const char* prefix) {
-   wchar* wPrefix = NULL;
-   if (prefix) {
-      if ((uint)prefix < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter prefix: 0x%p (not a valid pointer)", prefix);
-      wPrefix = ansiToUtf16(prefix);
-   }
-   BOOL result = EnumWindowPropertiesW(hWnd, wPrefix);
-   if (wPrefix) free(wPrefix);
-   return result;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Enumerates the properties of the specified window.
- *
- * @param  HWND   hWnd - window handle
- * @param  wchar* name - limiting name prefix of properties to print
- *
- * @return BOOL - success status; FALSE if the function did not find any window properties or in case of errors
- */
-BOOL WINAPI EnumWindowPropertiesW(HWND hWnd, const wchar* prefix) {
-   if (prefix) {
-      if ((uint)prefix < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter prefix: 0x%p (not a valid pointer)", prefix);
-   }
-   int result = EnumPropsExW(hWnd, EnumWindowPropertiesProcW, (LPARAM)prefix);
-   return (result != -1);
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return a named integer value linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - integer name
- *
- * @return int - stored value or NULL if the name was not found or in case of errors
- */
-int WINAPI GetWindowIntegerA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   IntegerMap::iterator result = g_intWndProperties.find(key);
-
-   if (result != g_intWndProperties.end()) {
-      return result->second;
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return a named double value linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - double name
- *
- * @return double - stored value or NULL if the name was not found or in case of errors
- */
-double WINAPI GetWindowDoubleA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   DoubleMap::iterator result = g_doubleWndProperties.find(key);
-
-   if (result != g_doubleWndProperties.end()) {
-      return result->second;
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return a named string linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - string identifier
- *
- * @return char* - stored string or a NULL pointer if the name was not found or in case of errors
- */
-const char* WINAPI GetWindowStringA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   StringMap::iterator result = g_stringWndProperties.find(key);
-
-   if (result != g_stringWndProperties.end()) {
-      return result->second.c_str();
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Store a named integer value and link it to the specified window.
- *
- * @param  HWND  hWnd  - window handle
- * @param  char* name  - integer name
- * @param  int   value - integer value
- *
- * @return BOOL - success status
- */
-BOOL WINAPI SetWindowIntegerA(HWND hWnd, const char* name, int value) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   g_intWndProperties[key] = value;
-
-   return TRUE;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Store a named double value and link it to the specified window.
- *
- * @param  HWND   hWnd  - window handle
- * @param  char*  name  - integer name
- * @param  double value - double value
- *
- * @return BOOL - success status
- */
-BOOL WINAPI SetWindowDoubleA(HWND hWnd, const char* name, double value) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   g_doubleWndProperties[key] = value;
-
-   return TRUE;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Store a named string and link it to the specified window.
- *
- * @param  HWND  hWnd  - window handle
- * @param  char* name  - string identifier
- * @param  char* value - string (must not be a NULL pointer)
- *
- * @return BOOL - success status
- */
-BOOL WINAPI SetWindowStringA(HWND hWnd, const char* name, const char* value) {
-   if (!IsWindow(hWnd))                 return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER)  return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                          return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-   if ((uint)value < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter value: 0x%p (not a valid pointer)", value);
-
-   string key = to_string(hWnd).append("|").append(name);
-   g_stringWndProperties[key] = value;
-
-   return TRUE;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Remove a named integer value linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - integer name
- *
- * @return int - removed value or NULL if the name was not found or in case of errors
- */
-int WINAPI RemoveWindowIntegerA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   IntegerMap::iterator result = g_intWndProperties.find(key);
-
-   if (result != g_intWndProperties.end()) {
-      int value = result->second;
-      g_intWndProperties.erase(result);
-      return value;
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Remove a named double value linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - double name
- *
- * @return double - removed value or NULL if the name was not found or in case of errors
- */
-double WINAPI RemoveWindowDoubleA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return !error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return !error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return !error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   DoubleMap::iterator result = g_doubleWndProperties.find(key);
-
-   if (result != g_doubleWndProperties.end()) {
-      double value = result->second;
-      g_doubleWndProperties.erase(result);
-      return value;
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Remove a named string linked to the specified window.
- *
- * @param  HWND  hWnd - window handle
- * @param  char* name - double name
- *
- * @return char* - removed string or a NULL pointer if the name was not found or in case of errors
- */
-char* WINAPI RemoveWindowStringA(HWND hWnd, const char* name) {
-   if (!IsWindow(hWnd))                return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: 0x%p (not a window)", hWnd);
-   if ((uint)name < MIN_VALID_POINTER) return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter name: 0x%p (not a valid pointer)", name);
-   if (!*name)                         return (char*)!error(ERR_INVALID_PARAMETER, "invalid parameter name: \"\" (empty)");
-
-   string key = to_string(hWnd).append("|").append(name);
-   StringMap::iterator result = g_stringWndProperties.find(key);
-
-   if (result != g_stringWndProperties.end()) {
-      string value = result->second;
-      g_stringWndProperties.erase(result);      // invalidates result and releases result->second
-      return sdup(value.c_str());               // caller must free()
-   }
-   return NULL;
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Release all stored window properties. Called from DLL::onProcessDetach() only.
- */
-void WINAPI ReleaseWindowProperties() {
-   g_intWndProperties.clear();
-   g_doubleWndProperties.clear();
-   g_stringWndProperties.clear();
 }
