+++ src/lib/tester.cpp
@@ -2,12 +2,11 @@
 #include "lib/conversion.h"
 #include "lib/datetime.h"
 #include "lib/file.h"
-#include "lib/helper.h"
 #include "lib/math.h"
 #include "lib/string.h"
 #include "lib/terminal.h"
 #include "lib/tester.h"
-#include "lib/win32.h"
+#include "lib/window.h"
 
 #include <fstream>
 #include <windowsx.h>
