--- src/lib/win32.cpp
@@ -1,138 +1,0 @@
-#include "expander.h"
-#include "lib/conversion.h"
-#include "lib/helper.h"
-#include "lib/string.h"
-#include "lib/win32.h"
-
-
-/**
- * Return the classname of the specified window. Simplified version of user32::GetClassNameA().
- *
- * @param  HWND hWnd - window handle
- *
- * @return char* - classname or a NULL pointer in case of errors
- */
-char* WINAPI GetClassNameA(HWND hWnd) {
-   string className = getClassNameA(hWnd);
-   return sdup(className.c_str());              // caller must free()
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the classname of the specified window. Simplified version of user32::GetClassNameA().
- *
- * @param  HWND hWnd - window handle
- *
- * @return string - classname or an empty string in case of errors
- */
-string WINAPI getClassNameA(HWND hWnd) {
-   char className[256];                         // max. possible value
-   if (!GetClassNameA(hWnd, className, countof(className))) {
-      return _empty_str(error(ERR_WIN32_ERROR + GetLastError(), "GetClassNameA(hWnd=%p)", hWnd));
-   }
-   return string(className);
-}
-
-
-/**
- * Return the classname of the specified window. Simplified version of user32::GetClassNameW().
- *
- * @param  HWND hWnd - window handle
- *
- * @return wchar* - classname or a NULL pointer in case of errors
- */
-wchar* WINAPI GetClassNameW(HWND hWnd) {
-   wstring className = getClassNameW(hWnd);
-   return wsdup(className.c_str());             // caller must free()
-   #pragma EXPANDER_EXPORT
-}
-
-
-/**
- * Return the classname of the specified window. Simplified version of user32::GetClassNameW().
- *
- * @param  HWND hWnd - window handle
- *
- * @return wstring - classname or an empty string in case of errors; call GetLastError() for details
- */
-wstring WINAPI getClassNameW(HWND hWnd) {
-   wchar className[256];                        // max. possible value
-   if (!GetClassNameW(hWnd, className, countof(className))) {
-      return _empty_wstr(error(ERR_WIN32_ERROR + GetLastError(), "GetClassNameW(hWnd=%p)", hWnd));
-   }
-   return wstring(className);
-}
-
-
-/**
- * Enumerate the child windows of a window and send details to the debug output.
- *
- * @param  HWND hWnd                 - window to process; if NULL all top-level windows are enumerated
- * @param  BOOL recursive [optional] - whether to enumerate childs recursively (default: no)
- *
- * @return int - number of reported child windows or EMPTY (-1) in case of errors
- */
-int WINAPI EnumChildWindowsToDebug(HWND hWnd, BOOL recursive/*=FALSE*/) {
-   HWND hWndDesktop = GetDesktopWindow();
-   if (!hWnd) hWnd = hWndDesktop;
-   if (!IsWindow(hWnd)) return _EMPTY(error(ERR_INVALID_PARAMETER, "invalid parameter hWnd: %p (not a valid window)", hWnd));
-
-   /**
-    * A local function as substitute for missing lambda support in C++03.
-    * On error a window was destroyed cross-thread. Such a window is reported as "gone".
-    *
-    * @return int - number of reported child windows
-    */
-   struct local {
-      static int ProcessWindow(HWND hWnd, HWND hWndParent, HWND hWndDesktop, BOOL isRoot, BOOL recursive, uint level) {
-         uint indentLevel = (level > 0) ? (level-1) : 0;
-         string spaces(indentLevel * 2, ' ');
-         char* marker = level ? "-> " : "";
-
-         SetLastError(NO_ERROR);
-         wstring wndClass = getClassNameW(hWnd);
-         if (GetLastError()) return _NULL(debug_raw("  %s%s%p: (gone)  [%s]", spaces.c_str(), marker, hWnd, ErrorToStrA(GetLastError())));
-
-         wstring wndTitle = getInternalWindowTextW(hWnd);
-         if (GetLastError()) return _NULL(debug_raw("  %s%s%p: (gone)  [%s]", spaces.c_str(), marker, hWnd, ErrorToStrA(GetLastError())));
-
-         const char* sType = "";
-         if      (hWnd == hWndDesktop)       sType = " (desktop)";
-         else if (hWndParent == hWndDesktop) sType = " (top-level)";
-
-         int ctrlId = GetDlgCtrlID(hWnd);
-         if (!ctrlId && GetLastError()) return _NULL(debug_raw("  %s%s%p: (gone)  [%s]", spaces.c_str(), marker, hWnd, ErrorToStrA(GetLastError())));
-
-         if (!hWndParent || hWndParent == hWndDesktop) {
-            ctrlId = 0;
-         }
-         char* sCtrlId = asformat(" (id %d)", ctrlId);
-
-         debug_raw("  %s%s%p: %S \"%S\"%s%s", spaces.c_str(), marker, hWnd, wndClass.c_str(), wndTitle.c_str(), sType, ctrlId ? sCtrlId : "");
-         free(sCtrlId);
-
-         // enumerate child windows
-         int count = !isRoot;
-         if (isRoot || recursive) {
-            HWND hWndNext = GetWindow(hWnd, GW_CHILD);      // no more error handling as the window was successfully reported
-            while (hWndNext != 0) {
-               count += local::ProcessWindow(hWndNext, hWnd, hWndDesktop, FALSE, recursive, level+1);
-               hWndNext = GetWindow(hWndNext, GW_HWNDNEXT);
-            }
-         }
-         return count;
-      }
-   };
-
-   // process the window
-   debug("for HWND %p", hWnd);
-   HWND hWndParent = GetAncestor(hWnd, GA_PARENT);
-   int count = local::ProcessWindow(hWnd, hWndParent, hWndDesktop, TRUE, recursive, 0);
-
-   // indicate if no child windows were found to visually distinguish from error conditions
-   if (!count) debug_raw("  -> (no child windows)");
-
-   return count;
-   #pragma EXPANDER_EXPORT
-}
